refactor: function organization and renaming
